사라지는 타일, 회전하면서 돌아다니는 오브젝트 수정

+++ Assets/Scripts/Object/MoveObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MoveObject : MonoBehaviour
 {
-    public int space;
+    public float space;
 
     public float speed;
